src/requirements.txt
anyio==3.7.1
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
dnspython==2.4.0
exceptiongroup==1.1.2
Flask==2.3.2
Flask-PyMongo==2.3.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
psycopg2==2.9.6
pymongo==4.4.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.31.0
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.3.6
zipp==3.16.2
